requirements.txt
langchain
langchain_google_genai 
google-generativeai
langchain-community
python-dotenv
pydantic
fastapi
uvicorn
